pyproject.toml
[tool.poetry]
name = "use-mysql"
version = "0.0.2"
description = ""
authors = ["miclon <jcnd@163.com>"]
readme = "README.md"
packages = [{ include = 'use_mysql', from = 'src' }]

[tool.poetry.dependencies]
python = "^3.8"
mysql-connector-python = "^8.1.0"


[tool.poetry.group.test.dependencies]
pylint = "*"
pytest = "*"
black = "*"
flake8 = "*"
isort = "*"
pre-commit = "*"
pre-commit-hooks = "*"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
